Minor changes to API. Minor changes to tests.

+++ include/ADT/dequeue.h
@@ -2,19 +2,20 @@
 #define DEQUEUE_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #define T dequeue_t
 typedef struct T *T;
 
-extern T     dequeue_new       (void);
-extern bool  dequeue_isempty   (T dequeue);
-extern int   dequeue_size      (T dequeue);
-extern void  dequeue_push_back (T dequeue, void *elem);
-extern void  dequeue_push_front(T dequeue, void *elem);
-extern void *dequeue_pop_back  (T dequeue);
-extern void *dequeue_pop_front (T dequeue);
-extern void  dequeue_clear     (T dequeue, void (*elem_free)(void *));
-extern void  dequeue_free      (T *dequeue, void (*elem_free)(void *));
+extern T          dequeue_new       (void);
+extern bool       dequeue_isempty   (T dequeue);
+extern uintmax_t  dequeue_size      (T dequeue);
+extern void       dequeue_push_back (T dequeue, void *elem);
+extern void       dequeue_push_front(T dequeue, void *elem);
+extern void      *dequeue_pop_back  (T dequeue);
+extern void      *dequeue_pop_front (T dequeue);
+extern void       dequeue_clear     (T dequeue, void (*elem_free)(void *));
+extern void       dequeue_free      (T *dequeue, void (*elem_free)(void *));
 
 #undef T
 #endif // DEQUEUE_H
